[hotfix] Refactor Kubernetes tests to make the test fixture reusable
