+++ flink-kubernetes/src/test/java/org/apache/flink/kubernetes/highavailability/KubernetesStateHandleStoreITCase.java
@@ -33,7 +33,6 @@ import org.junit.Test;
 
 import java.util.UUID;
 
-import static org.apache.flink.kubernetes.highavailability.KubernetesHighAvailabilityTestBase.LEADER_CONFIGMAP_NAME;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
@@ -47,6 +46,7 @@ import static org.hamcrest.Matchers.notNullValue;
  */
 public class KubernetesStateHandleStoreITCase extends TestLogger {
 
+    private static final String LEADER_CONFIGMAP_NAME = "leader-test-cluster";
     @ClassRule public static KubernetesResource kubernetesResource = new KubernetesResource();
 
     private final FlinkKubeClientFactory kubeClientFactory = new FlinkKubeClientFactory();
